\name{roblm}
\alias{fitted.roblm}
\alias{residuals.roblm}
\alias{coef.roblm}
\alias{vcov.roblm}
\title{Utility functions for weighted BACON linear regression}
\usage{
\method{fitted}{roblm}(object, ...)
\method{residuals}{roblm}(object, ...)
\method{coef}{roblm}(object, ...) 
\method{vcov}{roblm}(object, ...)
}
\arguments{
	\item{object}{object of class \code{roblm}.}
	\item{x}{object of class \code{roblm}.}
	\item{...}{additional arguments passed to the method.}
}
\description{
Utility functions
}
\seealso{
\code{\link{wBACON_reg}}, \code{\link[=summary.roblm]{summary}}.
}
