\name{wbacon-package}
\alias{wbacon-package}
\alias{wbacon}
\docType{package}
\title{Weighted BACON Algorithms for Multivariate Outlier Nomination (Detection) and Robust Linear Regression}
\description{
The package \pkg{wbacon} implements the BACON algorithms of Billor et al.
(2000) and some of the extensions proposed by Béguin and Hulliger (2008).
}
\details{
See \code{\link{wBACON}} to learn more on the BACON method for multivariate
outlier nomination (detection).

See \code{\link{wBACON_reg}} to learn more on the BACON method for robust
linear regression.
}
\author{
Tobias Schoch
}
\references{
Billor N., Hadi A.S. and Vellemann P.F. (2000). BACON: Blocked Adaptive
	Computationally efficient Outlier Nominators.
    \emph{Computational Statistics and Data Analysis}, \bold{34}, pp. 279--298.
    \doi{10.1016/S0167-9473(99)00101-2}

Béguin C. and Hulliger B. (2008). The BACON-EEM Algorithm for Multivariate
	Outlier Detection in Incomplete Survey Data. \emph{Survey Methodology},
	\bold{34}, pp. 91--103.
    \url{https://www150.statcan.gc.ca/n1/en/catalogue/12-001-X200800110616}

Schoch, T. (2021). wbacon: Weighted BACON algorithms for multivariate
    outlier nomination (detection) and robust linear regression,
    \emph{Journal of Open Source Software}, \bold{6} (62), 3238
    \doi{10.21105/joss.03238}
}
