\name{wbacon-package}
\alias{wbacon-package}
\alias{wbacon}
\docType{package}
\title{Weighted BACON Algorithms for Multivariate Outlier Nomination (Detection) and Robust Linear Regression}
\description{
The package \pkg{wbacon} implements the BACON algorithms of Billor et al.
(2000) and some of the extensions proposed by Béguin and Hulliger (2008).
}
\details{
See \code{\link{wBACON}} to learn more on the BACON method for multivariate
outlier nomination (detection).

See \code{\link{wBACON_reg}} to learn more on the BACON method for robust
linear regression.
}
\author{
Tobias Schoch
}
\references{
Billor N., Hadi A.S., Vellemann P.F. (2000). BACON: Blocked Adaptive
	Computationally efficient Outlier Nominators. \emph{Computational
	Statistics and Data Analysis} 34, pp. 279-298.

Béguin C., Hulliger B. (2008). The BACON-EEM Algorithm for Multivariate
	Outlier Detection in Incomplete Survey Data. \emph{Survey Methodology}
	34, pp. 91-103.
}
